Move the app adapter into the launcher package
