+++ workshops/InversionOfControl/default/app/src/main/java/com/novoda/iockata/launcher/AppAdapter.kt
@@ -1,4 +1,4 @@
-package com.novoda.iockata.apps
+package com.novoda.iockata.launcher
 
 import android.app.Activity
 import android.support.v7.widget.RecyclerView
@@ -6,11 +6,12 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import android.widget.ImageView
 import com.novoda.iockata.R
-import com.novoda.iockata.core.devicefeatures.DeviceFeature
+import com.novoda.iockata.apps.App
 import com.novoda.iockata.apps.browser.BrowserApp
 import com.novoda.iockata.apps.camera.CameraApp
 import com.novoda.iockata.apps.phone.PhoneApp
 import com.novoda.iockata.apps.toaster.ToasterApp
+import com.novoda.iockata.core.devicefeatures.DeviceFeature
 
 class AppAdapter(activity: Activity, deviceFeature: DeviceFeature) : RecyclerView.Adapter<AppAdapter.ViewHolder>() {
